+++ content/!del/geometry/HullDiameter.h
@@ -11,10 +11,10 @@
 #pragma once
 #include "Point.h"
 
-typedef Point<ll> P;
+typedef Point<int> P;
 array<P, 2> hullDiameter(vector<P> S) {
 	int n = sz(S), j = n < 2 ? 0 : 1;
-	pair<ll, array<P, 2>> res({0, {S[0], S[0]}});
+	pair<int, array<P, 2>> res({0, {S[0], S[0]}});
 	rep(i,0,j)
 		for (;; j = (j + 1) % n) {
 			res = max(res, {(S[i] - S[j]).dist2(), {S[i], S[j]}});
