+++ content/!del/numerical/BerlekampMassey.h
@@ -16,23 +16,23 @@
 
 #include "../number-theory/ModPow.h"
 
-vector<ll> berlekampMassey(vector<ll> s) {
+vector<int> berlekampMassey(vector<int> s) {
 	int n = sz(s), L = 0, m = 0;
-	vector<ll> C(n), B(n), T;
+	vector<int> C(n), B(n), T;
 	C[0] = B[0] = 1;
 
-	ll b = 1;
+	int b = 1;
 	rep(i,0,n) { ++m;
-		ll d = s[i] % mod;
+		int d = s[i] % mod;
 		rep(j,1,L+1) d = (d + C[j] * s[i - j]) % mod;
 		if (!d) continue;
-		T = C; ll coef = d * modpow(b, mod-2) % mod;
+		T = C; int coef = d * modpow(b, mod-2) % mod;
 		rep(j,m,n) C[j] = (C[j] - coef * B[j - m]) % mod;
 		if (2 * L > i) continue;
 		L = i + 1 - L; B = T; b = d; m = 0;
 	}
 
 	C.resize(L + 1); C.erase(C.begin());
-	for (ll& x : C) x = (mod - x) % mod;
+	for (int& x : C) x = (mod - x) % mod;
 	return C;
 }
