+++ content/!del/numerical/NumberTheoreticTransform.h
@@ -18,16 +18,16 @@
 
 #include "../number-theory/ModPow.h"
 
-const ll mod = (119 << 23) + 1, root = 62; // = 998244353
+const int mod = (119 << 23) + 1, root = 62; // = 998244353
 // For p < 2^30 there is also e.g. 5 << 25, 7 << 26, 479 << 21
 // and 483 << 21 (same root). The last two are > 10^9.
-typedef vector<ll> vl;
+typedef vector<int> vl;
 void ntt(vl &a) {
 	int n = sz(a), L = 31 - __builtin_clz(n);
 	static vl rt(2, 1);
 	for (static int k = 2, s = 2; k < n; k *= 2, s++) {
 		rt.resize(n);
-		ll z[] = {1, modpow(root, mod >> s)};
+		int z[] = {1, modpow(root, mod >> s)};
 		rep(i,k,2*k) rt[i] = rt[i / 2] * z[i & 1] % mod;
 	}
 	vi rev(n);
@@ -35,7 +35,7 @@ void ntt(vl &a) {
 	rep(i,0,n) if (i < rev[i]) swap(a[i], a[rev[i]]);
 	for (int k = 1; k < n; k *= 2)
 		for (int i = 0; i < n; i += 2 * k) rep(j,0,k) {
-			ll z = rt[j + k] * a[i + j + k] % mod, &ai = a[i + j];
+			int z = rt[j + k] * a[i + j + k] % mod, &ai = a[i + j];
 			a[i + j + k] = ai - z + (z > ai ? mod : 0);
 			ai += (ai + z >= mod ? z - mod : z);
 		}
@@ -49,7 +49,7 @@ vl conv(const vl &a, const vl &b) {
 	L.resize(n), R.resize(n);
 	ntt(L), ntt(R);
 	rep(i,0,n)
-		out[-i & (n - 1)] = (ll)L[i] * R[i] % mod * inv % mod;
+		out[-i & (n - 1)] = (int)L[i] * R[i] % mod * inv % mod;
 	ntt(out);
 	return {out.begin(), out.begin() + s};
 }
