+++ content/number-theory/CRT.h
@@ -14,9 +14,9 @@
 
 #include "euclid.h"
 
-ll crt(ll a, ll m, ll b, ll n) {
+int crt(int a, int m, int b, int n) {
 	if (n > m) swap(a, b), swap(m, n);
-	ll x, y, g = euclid(m, n, x, y);
+	int x, y, g = euclid(m, n, x, y);
 	assert((a - b) % g == 0); // else no solution
 	x = (b - a) % n * x % n / g * m + a;
 	return x < 0 ? x + m*n/g : x;
