+++ content/number-theory/MillerRabin.h
@@ -12,12 +12,12 @@
 
 #include "ModMulLL.h"
 
-bool isPrime(ull n) {
+bool isPrime(uint n) {
 	if (n < 2 || n % 6 % 4 != 1) return (n | 1) == 3;
-	ull A[] = {2, 325, 9375, 28178, 450775, 9780504, 1795265022},
+	uint A[] = {2, 325, 9375, 28178, 450775, 9780504, 1795265022},
 	    s = __builtin_ctzll(n-1), d = n >> s;
-	for (ull a : A) {   // ^ count trailing zeroes
-		ull p = modpow(a%n, d, n), i = s;
+	for (uint a : A) {   // ^ count trailing zeroes
+		uint p = modpow(a%n, d, n), i = s;
 		while (p != 1 && p != n - 1 && a % n && i--)
 			p = modmul(p, p, n);
 		if (p != n-1 && i != s) return 0;
