+++ content/number-theory/ModularArithmetic.h
@@ -10,19 +10,19 @@
 
 #include "euclid.h"
 
-const ll mod = 17; // change to something else
+const int mod = 17; // change to something else
 struct Mod {
-	ll x;
-	Mod(ll xx) : x(xx) {}
+	int x;
+	Mod(int xx) : x(xx) {}
 	Mod operator+(Mod b) { return Mod((x + b.x) % mod); }
 	Mod operator-(Mod b) { return Mod((x - b.x + mod) % mod); }
 	Mod operator*(Mod b) { return Mod((x * b.x) % mod); }
 	Mod operator/(Mod b) { return *this * invert(b); }
 	Mod invert(Mod a) {
-		ll x, y, g = euclid(a.x, mod, x, y);
+		int x, y, g = euclid(a.x, mod, x, y);
 		assert(g == 1); return Mod((x + mod) % mod);
 	}
-	Mod operator^(ll e) {
+	Mod operator^(int e) {
 		if (!e) return Mod(1);
 		Mod r = *this ^ (e / 2); r = r * r;
 		return e&1 ? *this * r : r;
