+++ content/number-theory/del/ContinuedFractions.h
@@ -15,11 +15,11 @@
  */
 
 typedef double d; // for N ~ 1e7; long double for N ~ 1e9
-pair<ll, ll> approximate(d x, ll N) {
-	ll LP = 0, LQ = 1, P = 1, Q = 0, inf = LLONG_MAX; d y = x;
+pair<int, int> approximate(d x, int N) {
+	int LP = 0, LQ = 1, P = 1, Q = 0, inf = LLONG_MAX; d y = x;
 	for (;;) {
-		ll lim = min(P ? (N-LP) / P : inf, Q ? (N-LQ) / Q : inf),
-		   a = (ll)floor(y), b = min(a, lim),
+		int lim = min(P ? (N-LP) / P : inf, Q ? (N-LQ) / Q : inf),
+		   a = (int)floor(y), b = min(a, lim),
 		   NP = b*P + LP, NQ = b*Q + LQ;
 		if (a > b) {
 			// If b > a/2, we have a semi-convergent that gives us a
