+++ content/number-theory/euclid.h
@@ -8,8 +8,8 @@
  */
 #pragma once
 
-ll euclid(ll a, ll b, ll &x, ll &y) {
+int euclid(int a, int b, int &x, int &y) {
 	if (!b) return x = 1, y = 0, a;
-	ll d = euclid(b, a % b, y, x);
+	int d = euclid(b, a % b, y, x);
 	return y -= a/b * x, d;
 }
