+++ content/various/FastMod.h
@@ -13,11 +13,11 @@
  */
 #pragma once
 
-typedef unsigned long long ull;
+typedef unsigned long long uint;
 struct FastMod {
-	ull b, m;
-	FastMod(ull b) : b(b), m(-1ULL / b) {}
-	ull reduce(ull a) { // a % b + (0 or b)
-		return a - (ull)((__uint128_t(m) * a) >> 64) * b;
+	uint b, m;
+	FastMod(uint b) : b(b), m(-1ULL / b) {}
+	uint reduce(uint a) { // a % b + (0 or b)
+		return a - (uint)((__uint128_t(m) * a) >> 64) * b;
 	}
 };
